Backend/requirements.txt
fastapi==0.118.2
pwdlib==0.2.1
pydantic==2.12.0
PyJWT==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
psycopg2==2.9.10
pwdlib[argon2]==0.2.1
uvicorn[standard]==0.37.0